pyproject.toml
[project]
name = "inter_exchange_arbitrage"
version = "0.1.0"
description = "A bot for inter-exchange arbitrage trading."
authors = [{ name = "You", email = "you@example.com" }]
dependencies = [
    "ccxt",
    "loguru",
    "PyYAML",
    "orjson",
]

[project.scripts]
arb-bot = "arbitrage_bot.cli:main"

[tool.uv]
dev-dependencies = [
    "pytest",
] 